edit-group: move into groups folder
This commit moves the controller for group editing
into the Groups folder.
